refactor: moved feature flags to its own module
