+++ cognite_toolkit/_cdf_tk/cdf_toml.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import sys
 from dataclasses import dataclass, field
-from pathlib import Path
 from typing import Any, ClassVar
 
 from cognite_toolkit import _version
