+++ tests/test_unit/test_cdf_tk/test_commands/test_feature_flags.py
@@ -1,4 +1,4 @@
-from cognite_toolkit._cdf_tk.commands.featureflag import FeatureFlag
+from cognite_toolkit._cdf_tk.feature_flags import FeatureFlag
 
 
 class TestFeatureCommand:
